Ajout dans d'une variable dans URL et paramètre de requête

+++ src/main/java/be/heh/api_rest/controller/StudentResponse.java
@@ -7,5 +7,5 @@ import jakarta.validation.constraints.Size;
 
 import java.time.LocalDate;
 
-public record StudentResponse(int id,String firstName, String lastName, String email, LocalDate dateOfBirth) {
+public record StudentResponse(Long id,String firstName, String lastName, String email, LocalDate dateOfBirth) {
 }
